+++ src/cmp/Layout/cmp/Panel/cmp/Symbols/Sym/cmp/Label/index.jsx
@@ -1,12 +1,15 @@
+/* eslint-disable no-unused-vars */
 import { useContext} from 'react';
-import Box from '@mui/material/Box';
+import { Box, Checkbox } from '@mui/material';
 
 import ctx from 'src/Context';
 import ACTIONS from 'src/reducer/actions';
 import useStyles from './../../styles';
 import useElementStyles from './styles';
 
-const Label = ({sym}) => {
+
+
+const Label = ({sym, checked, onClick}) => {
     const classes = useStyles(),
         localCasses = useElementStyles(),
         {  dispatch} = useContext(ctx),
@@ -20,6 +23,7 @@ const Label = ({sym}) => {
     return <div className={localCasses.SectionLabel}>
         <div>
             <Box className={classes.Box}>
+                <Checkbox checked={checked} onClick={onClick}/>
                 <strong>Label: </strong><input type="text" onInput={updateLabel} size="small" label="Symbol label"  value={sym.label} />
             </Box>
         </div>